refactor: review and group scattered palindrome implementation (#7487)
* feat: add utility to find lowest base for palindromic numbers and implement palindrome check for singly linked lists

* feat: add palindrome check for linked lists and find lowest base palindrome implementation

* feat: add utility to check if an integer's binary representation is a palindrome

* feat: add algorithm to check if a singly linked list is a palindrome

* feat: add PalindromeNumber utility class to check for palindromic integers

* style: remove redundant import to fix checkstyle error

* style: add blank line between static and non-static imports
