+++ src/main/java/com/thealgorithms/bitmanipulation/BinaryPalindromeCheck.java
@@ -9,6 +9,12 @@ package com.thealgorithms.bitmanipulation;
  * </p>
  *
  * @author Hardvan
+ * @see com.thealgorithms.strings.Palindrome
+ * @see com.thealgorithms.stacks.PalindromeWithStack
+ * @see com.thealgorithms.maths.LowestBasePalindrome
+ * @see com.thealgorithms.datastructures.lists.PalindromeSinglyLinkedList
+ * @see com.thealgorithms.maths.PalindromePrime
+ * @see com.thealgorithms.maths.PalindromeNumber
  */
 public final class BinaryPalindromeCheck {
     private BinaryPalindromeCheck() {
