+++ src/test/java/com/thealgorithms/datastructures/lists/PalindromeSinglyLinkedListTest.java
@@ -1,9 +1,8 @@
-package com.thealgorithms.misc;
+package com.thealgorithms.datastructures.lists;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.thealgorithms.datastructures.lists.SinglyLinkedList;
 import org.junit.jupiter.api.Test;
 
 public class PalindromeSinglyLinkedListTest {
