+++ src/test/java/com/thealgorithms/maths/LowestBasePalindromeTest.java
@@ -1,4 +1,4 @@
-package com.thealgorithms.others;
+package com.thealgorithms.maths;
 
 import java.util.ArrayList;
 import java.util.Arrays;
